🛠️ refactor: consolidate memory types by relocating them from shared CLI directory to core memory package
