+++ src/my-memory/core/memory-service.ts
@@ -21,7 +21,7 @@ import { writeProjectMap } from '../tier1-map/write-project-map';
 import { syncAllInstructionFiles, syncInstructionFileForSlug } from '../tier1-map/sync-instructions';
 import { detectToolchain, installToolchain, type ProgressFn, type ToolchainStatus } from '../tier2-graph/toolchain';
 import { ensureGraphifyOutIgnored, runGraphify } from '../tier2-graph/graphify-runner';
-import type { MemorySnapshot, MemoryBuildResult } from '../../my-cli/shared/memory-types';
+import type { MemorySnapshot, MemoryBuildResult } from '../memory-types';
 
 export class MemoryService {
 	private enabled = false;
