+++ src/my-memory/memory-types.ts
